[Admin] Introduce CompositeNotificationProvider + extract GUS notifications
