+++ src/Sylius/Bundle/AdminBundle/Notification/NotificationProviderInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\AdminBundle\Notification;
+
+interface NotificationProviderInterface
+{
+    /** @return array<array-key, mixed> */
+    public function getNotifications(): array;
+}
